+++ chronograf.go
@@ -137,7 +137,7 @@ type Response interface {
 
 // Source is connection information to a time-series data store.
 type Source struct {
-	ID                 int    `json:"id,omitempty,string"`          // ID is the unique ID of the source
+	ID                 int    `json:"id,string"`                    // ID is the unique ID of the source
 	Name               string `json:"name"`                         // Name is the user-defined name for the source
 	Type               string `json:"type,omitempty"`               // Type specifies which kinds of source (enterprise vs oss)
 	Username           string `json:"username,omitempty"`           // Username is the username to connect to the source