+++ ui/src/admin/components/UserRow.js
@@ -6,9 +6,10 @@ import UserEditingRow from 'src/admin/components/UserEditingRow'
 import MultiSelectDropdown from 'shared/components/MultiSelectDropdown'
 import ConfirmButtons from 'shared/components/ConfirmButtons'
 import DeleteConfirmTableCell from 'shared/components/DeleteConfirmTableCell'
+import ChangePassRow from 'src/admin/components/ChangePassRow'
 
 const UserRow = ({
-  user: {name, roles, permissions},
+  user: {name, roles, permissions, password},
   user,
   allRoles,
   allPermissions,
@@ -21,6 +22,7 @@ const UserRow = ({
   onDelete,
   onUpdatePermissions,
   onUpdateRoles,
+  onUpdatePassword,
 }) => {
   const handleUpdatePermissions = (allowed) => {
     onUpdatePermissions(user, [{scope: 'all', allowed}])
@@ -28,6 +30,10 @@ const UserRow = ({
 
   const handleUpdateRoles = (roleNames) => {
     onUpdateRoles(user, allRoles.filter(r => roleNames.find(rn => rn === r.name)))
+  }
+
+  const handleUpdatePassword = () => {
+    onUpdatePassword(user, password)
   }
 
   if (isEditing) {
@@ -69,6 +75,9 @@ const UserRow = ({
             /> : null
         }
       </td>
+      <td className="text-right" style={{width: "300px"}}>
+        <ChangePassRow onEdit={onEdit} onApply={handleUpdatePassword} user={user} />
+      </td>
       <DeleteConfirmTableCell onDelete={onDelete} item={user} />
     </tr>
   )
@@ -91,6 +100,7 @@ UserRow.propTypes = {
     permissions: arrayOf(shape({
       name: string,
     })),
+    password: string,
   }).isRequired,
   allRoles: arrayOf(shape()),
   allPermissions: arrayOf(string),
@@ -103,6 +113,7 @@ UserRow.propTypes = {
   onDelete: func.isRequired,
   onUpdatePermissions: func,
   onUpdateRoles: func,
+  onUpdatePassword: func,
 }
 
 export default UserRow